refactor: move UserRepository interface to domain layer
